Reduce scope of functions
