+++ src/ripple/unity/sslutil.cpp
@@ -24,4 +24,3 @@
 #include <ripple/sslutil/impl/CBigNum.cpp>
 #include <ripple/sslutil/impl/ECDSACanonical.cpp>
 #include <ripple/sslutil/impl/DHUtil.cpp>
-#include <ripple/sslutil/impl/HashUtilities.cpp>
